+++ docs/chapter_blurbs.md
@@ -35,7 +35,7 @@ What does each one do?
 - **ch33_pitch**: Defines Pitch Tools for gifts, World Scenarios.
 - **ch34_finance**: Defines Finance Tools for 
 - **ch35_person_viewer**: Tools for Visualizing PersonUnits
-- **ch50_world_app**: ETL app to create day punches from excel files.
+- **ch90_world_app**: ETL app to create day punches from excel files.
 - **ch97_docs_builder**: Defines Tools that create documentation.
 - **ch98_linter**: Linter for repo.
 - **ch99_glossary**: Where keg terms will be defined, described, ranked.
