+++ docs/keywords_by_chapter.md
@@ -33,7 +33,7 @@
 - ch33_pitch: Deal, Gift, Peer, PitchID, PitchUnit, ch33, deal, gift, offer, peer, pitch, pitch_id, pitcher_name, pitchs, request
 - ch34_finance: ch34, finance, financial
 - ch35_person_viewer: ch35
-- ch50_world_app: ch50
+- ch90_world_app: ch90
 - ch97_docs_builder: ch97
 - ch98_linter: ch98
 - ch99_glossary: ch99
