+++ src/ch90_world_app/_ref/ch90_ref.json
@@ -1,7 +1,7 @@
 {
   "chapter_blurb": "ETL app to create day punches from excel files.", 
   "chapter_content": "Simple ETL app to create a world database from excel files.", 
-  "chapter_description": "ch50_world_app", 
-  "chapter_number": 50, 
+  "chapter_description": "ch90_world_app", 
+  "chapter_number": 90, 
   "ontology summary": "The entry point for most excel savy users."
 }
