+++ src/ch90_world_app/test/test_z_app_runs.py
@@ -1,4 +1,4 @@
-from ch50_world_app.w1_app import ETLApp
+from ch90_world_app.w1_app import ETLApp
 from os import environ as os_environ
 from pytest import fail as pytest_fail
