+++ src/ch99_glossary/derived/question_tier.json
@@ -44,7 +44,7 @@
     "ch33": {"keg_rank": 42, "question_tier": 6, "chs": "33:"},
     "ch34": {"keg_rank": 43, "question_tier": 6, "chs": "34:"},
     "ch35": {"keg_rank": 44, "question_tier": 6, "chs": "35:"},
-    "ch50": {"keg_rank": 45, "question_tier": 6, "chs": "36:"},
+    "ch90": {"keg_rank": 45, "question_tier": 6, "chs": "36:"},
     "ch97": {"keg_rank": 46, "question_tier": 6, "chs": "97:"},
     "ch98": {"keg_rank": 47, "question_tier": 6, "chs": "98:"},
     "ch99": {"keg_rank": 48, "question_tier": 6, "chs": "99:"},
