+++ src/ch99_glossary/keywords_src.json
@@ -141,7 +141,7 @@
   "ch33": {"valid_ch": "33:"}, 
   "ch34": {"valid_ch": "34:"}, 
   "ch35": {"valid_ch": "35:"}, 
-  "ch50": {"valid_ch": "36:"}, 
+  "ch90": {"valid_ch": "36:"}, 
   "ch97": {"valid_ch": "97:"}, 
   "ch98": {"valid_ch": "98:"}, 
   "ch99": {"valid_ch": "99:"},
